activities user

+++ src/posts/posts.service.ts
@@ -1,10 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Type } from '@nestjs/common';
 import { PublishPosts } from './publish-posts.schema';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model, Types } from 'mongoose';
 import { FilesService } from 'src/files/files.service';
-import { NOTIFICATION_EVENT, PRIVACY } from 'src/enum/enum';
-import { AddCommentDto, AddMarkPostDto, AddRepostDto, GetMarkPostDto, GetPostDto, GetPostsDto } from './posts.dto';
+import { NOTIFICATION_EVENT, NOTIFICATION_POST, PRIVACY } from 'src/enum/enum';
+import { AddCommentDto, AddMarkPostDto, AddRepostDto, ChangePostPrivacyDto, DeletePostDto, GetMarkPostDto, GetPostDto, GetPostsDto, NotificationPostDto, PostHideDto, UpdatePinPostDto } from './posts.dto';
 import { Likes } from 'src/likes/likes.schema';
 import { PublishComments } from './publish-comments.schema';
 import { UserIdentity } from 'src/user-identity/user-identity.schema';
@@ -12,6 +12,11 @@ import { NotificationService } from 'src/notification/notification.service';
 import { Repost } from './repost.schema';
 import { MarkPost } from './posts-mark.schema';
 import { IDUserDto } from 'src/user/user.dto';
+import { PostNotification } from './post-notification.schema';
+import { User } from 'src/user/user.schema';
+import { PostPin } from './post-pin.schema';
+import { UserIdDTO } from 'src/notification/notification.dto';
+import { PostHide } from './post-hide.schema';
 
 @Injectable()
 export class PostsService {
@@ -19,8 +24,10 @@ export class PostsService {
     constructor(
         @InjectModel(PublishPosts.name)
         private readonly publishPostsModel: Model<PublishPosts>,
+
         @InjectModel(Likes.name)
         private readonly likesModel: Model<Likes>,
+
         @InjectModel(PublishComments.name)
         private readonly publishCommentsModel: Model<PublishComments>,
 
@@ -30,9 +37,18 @@ export class PostsService {
         @InjectModel(MarkPost.name)
         private readonly markPostModel: Model<MarkPost>,
 
+        @InjectModel(PostNotification.name)
+        private readonly postNotificationModel: Model<PostNotification>,
 
-        @InjectModel(UserIdentity.name)
-        private readonly userIdentity: Model<UserIdentity>,
+        @InjectModel(User.name)
+        private readonly userModel: Model<User>,
+
+        @InjectModel(PostPin.name)
+        private readonly postPinModel: Model<PostPin>,
+
+        @InjectModel(PostHide.name)
+        private readonly postHideModel: Model<PostHide>,
+        
         private filesService: FilesService,
         private notificationService: NotificationService
     ) { }
@@ -46,26 +62,62 @@ export class PostsService {
             const skip = (body.pageNumber - 1) * pageSize;
             let allPosts = []
 
-            allPosts = await this.repostModel.find()
-                .skip(skip)
-                .limit(pageSize)
-                .populate({
-                    path: 'postId',
-                    populate: {
-                        path: 'userId likes',
-                        select: 'fullName avatarFileName usersId',
-                    },
+            const hideList = await this.postHideModel.find({ownerId: new Types.ObjectId(userId)})
 
-                })
-                .populate({
-                    path: 'repostedUserId',
-                    select: 'fullName avatarFileName',
-                })
-                .sort({ createdRepostDate: -1 })
+            console.log(userId,"hideList",hideList);
+            
+            const hideListRepost =  hideList.filter(item => !Boolean(item?.hideUserId)).map(item => item?.hideRepostId)
+            const hideListOwnerPost = hideList.filter(item => !Boolean(item?.hideRepostId)).map(item => item?.hideUserId)
+
+            console.log(hideListRepost, hideListOwnerPost);
+            
+
+            if (body?.listPinedPost?.length) {
+                const listPinnedPost = body.listPinedPost.map(item => new Types.ObjectId(item))
+                allPosts = await this.repostModel.find({ _id: { $in: listPinnedPost } })
+                    .populate({
+                        path: 'postId',
+                        populate: {
+                            path: 'userId likes',
+                            select: 'fullName avatarFileName usersId',
+                        },
+
+                    })
+                    .populate({
+                        path: 'repostedUserId',
+                        select: 'fullName avatarFileName',
+                    });
+            }
+            else {
+                allPosts = await this.repostModel.find(
+                    {
+                        $and: [
+                             { _id: { $nin: hideListRepost } },
+                            { ownerId: { $nin: hideListOwnerPost } },
+                        ],
+                    }                    
+                )
+                    .skip(skip)
+                    .limit(pageSize)
+                    .populate({
+                        path: 'postId',
+                        populate: {
+                            path: 'userId likes',
+                            select: 'fullName avatarFileName usersId',
+                        },
+
+                    })
+                    .populate({
+                        path: 'repostedUserId',
+                        select: 'fullName avatarFileName',
+                    })
+                    .sort({ createdRepostDate: -1 })
+                    .exec()
+            }
 
             if (body?.isMarkedOption) {
                 const arrMarkPostId = await this.getPostByMark({ marckedUserId: userId })
-             
+
                 allPosts = [...new Set(allPosts.map(item => item.postId?._id.toString()))].map(id => allPosts.find(post => post.postId?._id.toString() === id));
                 allPosts = allPosts.filter(item => {
                     return arrMarkPostId.includes(item.postId?._id.toString());
@@ -85,7 +137,31 @@ export class PostsService {
                         ]
                     })
 
+
+                    const isNotificatedComment = await this.postNotificationModel.findOne({
+                        $and: [
+                            { postId: new Types.ObjectId(post.postId._id) },
+                            { userId: body.userId },
+                            { typeNotification: NOTIFICATION_POST.COMMENT }
+                        ]
+                    })
+
+                    const isNotificatedPost = await this.postNotificationModel.findOne({
+                        $and: [
+                            { postId: new Types.ObjectId(post.postId._id) },
+                            { userId: body.userId },
+                            { typeNotification: NOTIFICATION_POST.POST }
+                        ]
+                    })
+
+                    const isPined = await this.postPinModel.findOne({
+                        $and: [
+                            { repostId: new Types.ObjectId(post._id) },
+                            { userId: new Types.ObjectId(body.userId) },
+                        ]
+                    })
                     return {
+                        repostId: post._id,
                         ...post?.postId?.toObject(),
                         repostedUserId: post?.repostedUserId ? post?.repostedUserId?.toObject() : null,
                         likeId: post?.postId?.toObject().likes?._id || '',
@@ -96,7 +172,10 @@ export class PostsService {
                         createdRepostDate: post.createdRepostDate,
                         isReposted: Boolean(isReposted),
                         isMarked: Boolean(isMarked),
-
+                        isNotificatedComment: Boolean(isNotificatedComment),
+                        isNotificatedPost: Boolean(isNotificatedPost),
+                        isPined: Boolean(isPined),
+                        isPinedPostFlag: Boolean(body?.listPinedPost?.length > 0)
                     }
                 }
                 )
@@ -134,7 +213,33 @@ export class PostsService {
                     { postId: new Types.ObjectId(postId) },
                 ]
             })
+
+            const isNotificatedComment = await this.postNotificationModel.findOne({
+                $and: [
+                    { postId: new Types.ObjectId(post._id) },
+                    { userId: body.userId },
+                    { typeNotification: NOTIFICATION_POST.COMMENT }
+                ]
+            })
+
+            const isNotificatedPost = await this.postNotificationModel.findOne({
+                $and: [
+                    { postId: new Types.ObjectId(post._id) },
+                    { userId: body.userId },
+                    { typeNotification: NOTIFICATION_POST.POST }
+                ]
+            })
+
+            
+            const isPined = await this.postPinModel.findOne({
+                $and: [
+                    { repostId: new Types.ObjectId(post._id) },
+                    { userId: new Types.ObjectId(body.userId) },
+                ]
+            })
+
             const postWithLike = {
+                repostId: post._id,
                 ...post.toObject(),
                 likeId: post.toObject().likes?._id || '',
                 likes: post.likes?.usersId.length || 0,
@@ -143,6 +248,11 @@ export class PostsService {
                 isMarked: Boolean(isMarked),
                 countComments,
                 countReposts,
+
+                isNotificatedComment: Boolean(isNotificatedComment),
+                isNotificatedPost: Boolean(isNotificatedPost),
+                isPined: Boolean(isPined),
+
             }
 
             return {
@@ -223,6 +333,15 @@ export class PostsService {
         }
     }
 
+    async getPostPin(body: UserIdDTO) {
+        try {
+            const userId = new Types.ObjectId(body.userId)
+            return await this.postPinModel.find({ userId: userId })
+        } catch (error) {
+            throw new error
+        }
+    }
+
     async addPost(
         { payload, files }: {
             payload:
@@ -233,11 +352,31 @@ export class PostsService {
                 text: string,
                 userId: string,
                 repostedUserId: string | null,
-                coordinates: { lat: number; lng: number }
+                coordinates: { lat: number; lng: number },
+                deletedFiles?: string[]
+                postId?: string
             }, files: Array<Express.Multer.File>
         }
     ) {
         try {
+
+            if (payload?.postId) {
+
+                const { text, title, coordinates, privacyPost } = payload
+                const newFilesName = await this.filesService.uploadFiles(files, 'uploads/publish_post', false)
+                const postId = new Types.ObjectId(payload?.postId)
+                const post = await this.publishPostsModel.findOne({ _id: postId })
+
+                const postFileName = post?.filesName.filter(item => !payload.deletedFiles.includes(item))
+
+                const filesName = [...newFilesName, ...postFileName]
+                await this.filesService.deleteFiles(payload?.deletedFiles, 'uploads/publish_post')
+
+                await post.updateOne({
+                    text, title, coordinates, filesName, privacyPost
+                })
+                return
+            }
 
             let repostedUserId = null
             const userId = new Types.ObjectId(payload.userId)
@@ -263,12 +402,45 @@ export class PostsService {
             }
 
             await this.repostModel.create({
-                repostedUserId, postId: post._id
+                repostedUserId, postId: post._id,
+                ownerId: userId
             })
 
             return post
         } catch (error) {
             throw new error
+        }
+    }
+
+    async changePostPrivacy(body: ChangePostPrivacyDto) {
+        try {
+            const postId = new Types.ObjectId(body.postId)
+            delete body.postId
+            await this.publishPostsModel.updateOne({ _id: postId }, { ...body })
+        } catch (error) {
+            throw new error
+        }
+    }
+
+    async updatePostPin(body: UpdatePinPostDto) {
+        try {
+            const repostId = new Types.ObjectId(body.repostId)
+            const userId = new Types.ObjectId(body.userId)
+            const isPined = await this.postPinModel.findOne({
+                repostId: repostId
+            })
+
+            if (isPined) {
+                await this.postPinModel.deleteOne({
+                    repostId: repostId
+                })
+                return
+            }
+
+            await this.postPinModel.create({ repostId, userId: userId })
+
+        } catch (error) {
+            throw new Error(error)
         }
     }
 
@@ -278,6 +450,11 @@ export class PostsService {
             const userIdentityId = new Types.ObjectId(body.userIdentityId)
             const likes = (await this.likesModel.create({}))._id
             const postId = new Types.ObjectId(body.postId)
+
+            const rooms = (await this.postNotificationModel.find({ $and: [{ postId }, { typeNotification: NOTIFICATION_POST.COMMENT }] })).map(item => item.userId)
+            const { avatarFileName, fullName } = await this.userModel.findOne({ _id: userId })
+            await this.notificationService.sendNotification({ rooms, ownerId: body.userId, title: body.text, fileName: avatarFileName, name: fullName, event: NOTIFICATION_EVENT.NOTIFICATION_NEWS })
+
             return await this.publishCommentsModel.create({
                 userId,
                 userIdentityId,
@@ -290,32 +467,62 @@ export class PostsService {
         }
     }
 
-    async addRepost(body: AddRepostDto) {
+    async updateNotification(body: NotificationPostDto) {
         try {
-            const repostedUserId = new Types.ObjectId(body.repostedUserId)
             const postId = new Types.ObjectId(body.postId)
+            const isNotificated = await this.postNotificationModel.findOne({
+                $and: [
+                    { postId },
+                    { userId: body.userId },
+                    { typeNotification: body.typeNotification }
+                ]
+            })
 
-            await this.repostModel.create({ repostedUserId, postId })
+            if (isNotificated) {
+                await this.postNotificationModel.deleteOne({
+                    $and: [
+                        { postId },
+                        { userId: body.userId },
+                        { typeNotification: body.typeNotification }
+                    ]
+                })
+                return
+            }
+
+            await this.postNotificationModel.create({ postId, userId: body.userId, typeNotification: body.typeNotification })
+
         } catch (error) {
-            throw new error
+            throw new Error(error)
         }
     }
 
-    async deleteRepost(body: AddRepostDto) {
+    async updateRepost(body: AddRepostDto) {
         try {
             const repostedUserId = new Types.ObjectId(body.repostedUserId)
             const postId = new Types.ObjectId(body.postId)
 
-            await this.repostModel.deleteOne({
+            const isReposted = await this.repostModel.findOne({
                 $and: [
                     { repostedUserId },
                     { postId },
                 ]
             })
+
+            if (isReposted) {
+                await this.repostModel.deleteOne({
+                    $and: [
+                        { repostedUserId },
+                        { postId },
+                    ]
+                })
+                return
+            }
+            await this.repostModel.create({ repostedUserId, postId, createdRepostDate: new Date() })
         } catch (error) {
             throw new error
         }
     }
+
 
     async addMark(body: AddMarkPostDto) {
         try {
@@ -340,4 +547,36 @@ export class PostsService {
         }
 
     }
+
+    async hidePost(body:PostHideDto){
+        try {            
+            await this.postHideModel.create({
+                createdHideDate: new Date(),
+                ownerId:  new Types.ObjectId(body.ownerId),
+                hideRepostId: body?.hideRepostId ? new Types.ObjectId(body?.hideRepostId) : null,
+                hideUserId:body?.hideUserId ? new Types.ObjectId(body?.hideUserId): null
+
+            })
+        } catch (error) {
+            
+        }
+    }
+
+
+    async deletePost(body: DeletePostDto) {
+        try {
+            const postId = new Types.ObjectId(body.postId)
+            const { filesName, likes } = await this.publishPostsModel.findOneAndRemove({ _id: postId })
+
+            await this.repostModel.deleteMany({ postId })
+            await this.filesService.deleteFiles(filesName, 'uploads/publish_post')
+            await this.likesModel.deleteMany({ _id: new Types.ObjectId(likes) })
+            await this.publishCommentsModel.deleteMany({ postId })
+            await this.markPostModel.deleteMany({ postId })
+            await this.postNotificationModel.deleteMany({ postId })
+        } catch (error) {
+            throw new error
+        }
+    }
+
 }
